pyproject.toml
[project]
name = "based-schizo-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "disnake>=2.10.1,<3.0.0",
    "python-dotenv>=1.0.1,<2.0.0",
    "requests>=2.32.3,<3.0.0",
]
